+++ .env.template
@@ -3,4 +3,7 @@ FLASK_APP=todo_app/app
 FLASK_ENV=development
 
 # Change the following values for local development.
-SECRET_KEY=secret-key
+TRELLO_KEY=
+TRELLO_TOKEN=
+TODO_LIST_ID=
+DONE_LIST_ID=
